[🎉FEATURE] 어드민 상품 등록, 수정

+++ src/main/java/com/supernova/fashionnova/cart/dto/CartResponseDto.java
@@ -8,9 +8,9 @@ public class CartResponseDto {
 
     private final List<CartItemDto> cartItemDtoList;
 
-    private final int totalPrice;
+    private final Long totalPrice;
 
-    public CartResponseDto(List<CartItemDto> cartItemDtoList, int totalPrice) {
+    public CartResponseDto(List<CartItemDto> cartItemDtoList, Long totalPrice) {
         this.cartItemDtoList = cartItemDtoList;
         this.totalPrice = totalPrice;
     }
